WIP: refactor process candidate identity handler
