+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/enums/ProcessIdentityType.java
@@ -1,10 +1,6 @@
 package uk.gov.di.ipv.core.library.enums;
 
 public enum ProcessIdentityType {
-    ALL,
-    TICF_ONLY,
-    COI,
-    STORE_IDENTITY,
     EXISTING,
     NEW,
     PENDING,
